Change input structure to overcome push limit
With previous changes, the preimage would exceed the 520 bytes push limit in the input script for 3 parties, which is the minimum required. Optimizing the script produced some useful size reductions but in general, if possible, would obfuscate the script increasing the potential for bugs.

As an alternative, the input structure was reviewed and the preimage is now passed in multiple chunks. This non only overcomes the push limit but also simplifies the overall script logic.